fix test, move out redux connect from CoreLayout to UserProfileWidgetContainer
